requirements.txt
flask==2.1.1
openai==0.11.3
requests==2.26.0
python-dotenv==0.19.1
